pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pythia"
version = "0.3.0"
description = "Self-hosted AI search engine — SearXNG + Ollama + Oracle AI Vector Search"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
    "textual>=0.80",
    "typer>=0.12",
    "oracledb>=2.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
    "sse-starlette>=2.0",
    "sentence-transformers>=2.0",
    "scrapling[fetchers]>=0.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
pythia = "pythia.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
